--- a/codecov.yml
+++ b/codecov.yml
@@ -1,2 +1,5 @@
 ignore:
   - "src/neon/**"  # We currently can't run coverage on arm so we're excluding it
+  - "examples/*" # examples
+  - "benches/*" # examples
+  - "simd-fuzz-target/*" # fuzz testing