--- a/src/avx2/deser.rs
+++ b/src/avx2/deser.rs
@@ -11,9 +11,18 @@ use crate::Deserializer;
 pub use crate::Result;
 
 impl<'de> Deserializer<'de> {
-    #[allow(clippy::if_not_else, mutable_transmutes, clippy::transmute_ptr_to_ptr)]
+    #[allow(
+        clippy::if_not_else,
+        mutable_transmutes,
+        clippy::transmute_ptr_to_ptr,
+        clippy::too_many_lines,
+        clippy::cast_ptr_alignment,
+        clippy::cast_possible_wrap,
+        clippy::if_not_else,
+        clippy::too_many_lines
+    )]
     #[cfg_attr(not(feature = "no-inline"), inline(always))]
-    pub fn parse_str_<'invoke>(
+    pub(crate) fn parse_str_<'invoke>(
         input: &'de [u8],
         buffer: &'invoke mut [u8],
         mut idx: usize,
@@ -35,17 +44,13 @@ impl<'de> Deserializer<'de> {
         loop {
             let v: __m256i = if src.len() >= src_i + 32 {
                 // This is safe since we ensure src is at least 32 wide
-                #[allow(clippy::cast_ptr_alignment)]
-                unsafe {
-                    _mm256_loadu_si256(src.as_ptr().add(src_i) as *const __m256i)
-                }
+                unsafe { _mm256_loadu_si256(src.as_ptr().add(src_i) as *const __m256i) }
             } else {
                 unsafe {
                     padding
                         .get_unchecked_mut(..src.len() - src_i)
                         .clone_from_slice(src.get_unchecked(src_i..));
                     // This is safe since we ensure src is at least 32 wide
-                    #[allow(clippy::cast_ptr_alignment)]
                     _mm256_loadu_si256(padding.as_ptr() as *const __m256i)
                 }
             };
@@ -55,11 +60,9 @@ impl<'de> Deserializer<'de> {
             let bs_bits: u32 = unsafe {
                 static_cast_u32!(_mm256_movemask_epi8(_mm256_cmpeq_epi8(
                     v,
-                    #[allow(clippy::cast_possible_wrap)]
                     _mm256_set1_epi8(b'\\' as i8)
                 )))
             };
-            #[allow(clippy::cast_possible_wrap)]
             let quote_mask = unsafe { _mm256_cmpeq_epi8(v, _mm256_set1_epi8(b'"' as i8)) };
             let quote_bits = unsafe { static_cast_u32!(_mm256_movemask_epi8(quote_mask)) };
             if (bs_bits.wrapping_sub(1) & quote_bits) != 0 {
@@ -101,40 +104,30 @@ impl<'de> Deserializer<'de> {
         let mut dst_i: usize = 0;
 
         // To be more conform with upstream
-        #[allow(clippy::if_not_else)]
         loop {
             let v: __m256i = if src.len() >= src_i + 32 {
                 // This is safe since we ensure src is at least 32 wide
-                #[allow(clippy::cast_ptr_alignment)]
-                unsafe {
-                    _mm256_loadu_si256(src.as_ptr().add(src_i) as *const __m256i)
-                }
+                unsafe { _mm256_loadu_si256(src.as_ptr().add(src_i) as *const __m256i) }
             } else {
                 unsafe {
                     padding
                         .get_unchecked_mut(..src.len() - src_i)
                         .clone_from_slice(src.get_unchecked(src_i..));
                     // This is safe since we ensure src is at least 32 wide
-                    #[allow(clippy::cast_ptr_alignment)]
                     _mm256_loadu_si256(padding.as_ptr() as *const __m256i)
                 }
             };
 
-            #[allow(clippy::cast_ptr_alignment)]
-            unsafe {
-                _mm256_storeu_si256(buffer.as_mut_ptr().add(dst_i) as *mut __m256i, v)
-            };
+            unsafe { _mm256_storeu_si256(buffer.as_mut_ptr().add(dst_i) as *mut __m256i, v) };
 
             // store to dest unconditionally - we can overwrite the bits we don't like
             // later
             let bs_bits: u32 = unsafe {
                 static_cast_u32!(_mm256_movemask_epi8(_mm256_cmpeq_epi8(
                     v,
-                    #[allow(clippy::cast_possible_wrap)]
                     _mm256_set1_epi8(b'\\' as i8)
                 )))
             };
-            #[allow(clippy::cast_possible_wrap)]
             let quote_mask = unsafe { _mm256_cmpeq_epi8(v, _mm256_set1_epi8(b'"' as i8)) };
             let quote_bits = unsafe { static_cast_u32!(_mm256_movemask_epi8(quote_mask)) };
             if (bs_bits.wrapping_sub(1) & quote_bits) != 0 {