--- a/src/error.rs
+++ b/src/error.rs
@@ -112,6 +112,8 @@ impl Error {
         }
     }
 }
+
+#[cfg_attr(tarpaulin, skip)]
 impl fmt::Display for Error {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         write!(