This is a work branch for the 0.2 release where we can introduce breaking changes for things we do not like in 0.1.

So far:

- [x] Add U64 type
- [x] Remove deprecated functions
- [x] Box objects to reduce enum size
- [x] Add object and array access to value trait for convenience 
- [x] Arch alignment