@Licenser 2019-10-23
Either works for me. We can change everything other then the outside facing API as much as we want :D.

The 0.2 branch so far changes 2 things: it removes some functions that were bad and marked as deprecated in 0.1 already and adds the U64 as a natively supported type.

Since users will never notice how the parsing code works under the hood it's not constrained by the same restrictions :D.

Even adding 128 bit support (since we probably would want to feature flag that for perf reasons?)  would not change the API just add to it (which reminds me, no harm to add as_u128 and as_i128 to the value trait :D