@sunnygleason 2019-10-23
how are we doing on benchmarks lately? I wonder, are there API changes or alternate entry points we could consider to increase parsing speed relative to native simdjson?