@Licenser 2019-10-23
We don't have many benchmarks when it comes to the API, and that is a interesting topic since it is where simd-json.rs deviates a lot from simdjson.

There are 4 parts of the 'lifetime' of data:

1) decoding / parsing
2) reading / evaluating
3) editing / mutating
4) encoding / rendering

upstream simdjson is extremely good at 1) but makes 2) harder and probably slower and 3) nearly impossible. 4) I suspect will be faster as well since mutations do not happen.

A part of the speedup in 1) is by sacrificing 2) and 3) - instead of creating a DOM (like we do) it uses a tape, which is hands down faster - no allocation, no calculation of hashes, a unfragmented memory area.

However when reading that sacrifices a lot of convenience, you can't "take the `4th` element of an array" or "get the key `foo` from a object" all let alone do so in a nested way. That all has to be hand rolled, for objects every key/value pair has to be checked until the right key is found, for an array each element has to be passed to find the `n`th one.

Mutation (3)) with that model is simply not possible without first creating some kind of mutable structure out of it (which either means parsing into an object (which we have from serde support) or creating a DOM (the same way we do on parsing.

simdjson, as a low level library lets you pick your battles and if for example you only ever read, is significantly faster. simd-json.rs is more opinionated towards end to end workflow and convenience. That said I think on with parsing to objects we are not significantly slower