@sunnygleason 2019-11-05
ok only about 800 lines left to review! will come back to it soon. this is looking nice, great work! 🍻