@Licenser 2019-11-05
you can look at the topic based PR's:
* https://github.com/simd-lite/simdjson-rs/pull/73
* https://github.com/simd-lite/simdjson-rs/pull/74

I create a pr towards the 0.2 branch for every change, just merge them every few days so they don't block as they're all dependant on each other.