@Licenser 2019-11-06
> The only other things I was wondering about were:
> 
> * would it be useful to specify what version of simdjson upstream this is tracking?
Yes! But we'd need to figure out what it is :joy:
> * arch alignments to template-ize code, would that still be useful? (while we're changing everything 😝 )
Yes! either towards 0.1 or 0.2 I'd love if we'd get there! But as always, provided it's not a perf hit - sometimes generic code is not as fast as specific code, but we got our benchmarks and perf tests to decide on that ;)
> 
> In any case, this looks really nice, the benchmark results sound outstanding!