@Licenser 2019-11-11
@sunnygleason lets do a quick planning on the 0.2 release, I think getting the "this is following simdjson vXXX in is a good goal to have. but we'd need to pick a version and then go over it to see how we track. Would you want to get the arch alignment changes into 0.2 as well? They kind of fit to that theme. Is there anything I'm missing for 0.2 we'd want