@sunnygleason 2019-11-19
@Licenser I like that Value work you're doing. I'm doing an investigation of turning SimdInput into a struct/trait with associated functions to reduce/standardize code in `stage1`. Should have something (hopefully interesting) in the next couple days. Warm up the benchmarks!  😄  🍻