@Licenser 2019-11-26
👍 absolutely, as long as there is stuff to be done we can wait :) put a not down when you're ready to go .- I want to avoid the everyone is waiting for someone else problem ;) so as soon as you're ready, I'm ready.