@Licenser 2019-12-01
Perf:

![image](https://user-images.githubusercontent.com/119093/69917333-eb964380-1464-11ea-95c3-e6963c036ffb.png)


## master
```
======= simd_json ======== parse|stringify ===== parse|stringify ====
data/canada.json         410 MB/s   370 MB/s   620 MB/s   280 MB/s
data/citm_catalog.json  1060 MB/s   320 MB/s  1410 MB/s   380 MB/s
data/twitter.json        900 MB/s   410 MB/s   900 MB/s   370 MB/s
data/log.json           1080 MB/s   720 MB/s  1080 MB/s   720 MB/s
```

## this
```
======= simd_json ======== parse|stringify ===== parse|stringify ====
data/canada.json         530 MB/s   430 MB/s   710 MB/s   320 MB/s
data/citm_catalog.json  1130 MB/s   480 MB/s  1550 MB/s   610 MB/s
data/twitter.json        920 MB/s   600 MB/s  1000 MB/s   650 MB/s
data/log.json           1080 MB/s  1080 MB/s  1080 MB/s  1080 MB/s
```