@Licenser 2019-12-01
0.2 release
* upstream tracking version
* generalisation of utf8parsing
* generalisation of stage1
* api update
* perf improvements
* tape parsing